move utils to parse_common
